Final refactor
This should hopefully be the final command buffer refactor for now. The idea now is that command buffer allocations are separated into frame and persistent. Frame allocations have implicit usage constraints because they may be freed on the next frame cycle. This means that consumers allocate on demand and don't store the buffer, and the work must be synchronized with the frame cycle. Persistent allocations need to be freed explicitly, but they can be freed whenever and from any thread. The main restriction for both of these methods is that the thread that allocated must be the only thread that writes on it.