Change with FramePools
Apparently I was confused and did not realize that vkResetCommandPool doesn't actually reset the memory to be reused. As such, the frame pools now keep track of a free list of command buffers that have been allocated which get redistributed each frame. This allows the same behavior as before where we can fire-and-forget allocate command buffers without worrying about which thread they are on and storing them.