Another large command buffer / internal refactor
Yet another refactor. I realized that there was no way in the API for a user to submit commands that weren't synced with the frame. In order to accomplish this, SubmissionHandler can now handle generic submissions and the API submits directly to this. The user can now additionally push and execute commands similarly to the internal vulkan api. Command pushing was also reworked to submit immediately, which means the need for a CommandQueue was gone. DeleteQueue already had a sort of 'finalizer' behavior, so I just used this instead to run callbacks when pushed commands are completed.